backend/requirements.txt
# Core Django
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication
djangorestframework-simplejwt==5.3.1

# Filtering and Search
django-filter==23.5

# Image Processing
Pillow==10.4.0

# QR Code Generation
qrcode==7.4.2

# Environment Variables
python-decouple==3.8

# Production Server
gunicorn==21.2.0

# PostgreSQL Support (optional)
psycopg2-binary==2.9.9

# API Documentation
drf-spectacular==0.27.0

# Excel Export/Import (optional)
openpyxl==3.1.2
xlsxwriter==3.1.9

# PDF Generation (optional)
reportlab==4.0.9
weasyprint==60.2

# Caching and Background Tasks (optional)
redis==5.0.1
celery==5.3.4

# Security
django-ratelimit==4.1.0

# Monitoring and Error Tracking (optional)
sentry-sdk==1.39.2

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Google Gemini AI
google-generativeai==0.8.3
